Update qnap-disk-spindown.md
Hi Alex, thanks for the clear writeup! In my case the QNAP switched around the SSD and one of the HDDs on me... This addition prevents that from happening. 
I've also decided to use the full script you linked to as that is a little more robust with checking if the RAID is being rebuilt etc., but your script definitely put me on the right path, thanks!